requirements.txt
opencv-python-headless==4.8.1.78
Pillow==10.4.0
PyYAML==5.3.1